Pulled out Gage, Sample, and Sensor
